fix: Update tests for dashboard consolidation and fix major import issues
Major test fixes implemented:
- Fixed import paths in cache tests (cache -> analysis module structure)
- Fixed removed dashboard import issues in integration tests
- Updated comprehensive dashboard integration tests for SocketIO API changes
- Fixed method name mismatches in productivity data sources
- Fixed data comparison logic in health data sources

✅ Fixed test suites:
- All comprehensive dashboard integration tests (17/17 passing)
- Most cache tests (imports fixed, some assertion tweaks needed)
- CLI integration tests
- Import compatibility tests

🔧 Remaining issues to address:
- Performance test expectations (response time thresholds)
- Some template/HTML content expectations
- Memory usage test calibration

This commit brings the test suite from failing during collection to mostly passing,
with key integration tests validating the consolidated dashboard system works correctly.
