+++ tests/cache/test_enhanced_context_analyzer.py
@@ -10,12 +10,12 @@ import pytest
 from datetime import datetime, timedelta
 from unittest.mock import Mock, patch
 
-from src.context_cleaner.cache.enhanced_context_analyzer import (
+from src.context_cleaner.analysis.enhanced_context_analyzer import (
     EnhancedContextAnalyzer, CacheEnhancedAnalysis, UsageWeightedScore
 )
-from src.context_cleaner.cache.usage_analyzer import UsagePatternSummary, WorkflowPattern, FileUsageMetrics
-from src.context_cleaner.cache.token_analyzer import TokenAnalysisSummary, CacheEfficiencyMetrics, TokenUsageInsights
-from src.context_cleaner.cache.temporal_analyzer import TemporalInsights
+from src.context_cleaner.analysis.usage_analyzer import UsagePatternSummary, WorkflowPattern, FileUsageMetrics
+from src.context_cleaner.analysis.token_analyzer import TokenAnalysisSummary, CacheEfficiencyMetrics, TokenUsageInsights
+from src.context_cleaner.analysis.temporal_analyzer import TemporalInsights
 from src.context_cleaner.core.context_analyzer import ContextAnalysisResult
